feat(psci): guest-facing PSCI over the HVC conduit
HvcCall now carries the full 32-bit SMCCC function ID (NOVA IDs
are value-stable, existing subscribers unaffected). components/
psci claims the whole PSCI range: VERSION/FEATURES/AFFINITY_INFO
answer inline, SYSTEM_OFF retires the calling VM, SYSTEM_RESET
rides the warm-reset path; CPU_ON/OFF wait for multi-vCPU. Pure
dispatch table host-tested.

+++ src/components/trap_handler/include/trap_handler/hvc.hpp
@@ -24,8 +24,10 @@ namespace nova {
 inline constexpr std::uint64_t kSmcccNotSupported = ~0ULL;
 
 struct HvcCall {
-  TrapContext*  ctx     = nullptr;
-  std::uint16_t func_id = 0;
+  TrapContext* ctx = nullptr;
+  // Full SMCCC function ID (x0 bits 31:0) — standard ranges like PSCI
+  // (0x8400_xxxx) need all 32 bits; NOVA's own IDs stay below 0x10000.
+  std::uint32_t func_id = 0;
   bool          handled = false;
 };
